main - added user friendship management and validation for remove functions

+++ Server/Relationships/Follow/FollowService.cs
@@ -1,4 +1,5 @@
 ﻿using UBB_SE_2024_Gaborment.Server.Relationships.Block;
+using UBB_SE_2024_Gaborment.Server.Mocks;
 
 namespace UBB_SE_2024_Gaborment.Server.Relationships.Follow
 {
@@ -7,11 +8,20 @@ namespace UBB_SE_2024_Gaborment.Server.Relationships.Follow
         //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
         private BlockRepository _blockRepository;
         private FollowRepository _followRepository;
+        private UserServiceMock _userServiceMock;
 
         public FollowService(BlockRepository blockRepository, FollowRepository followRepository)
         {
             _blockRepository = blockRepository;
             _followRepository = followRepository;
+            _userServiceMock = new UserServiceMock();
+        }
+
+        public FollowService(BlockRepository blockRepository, FollowRepository followRepository, UserServiceMock userServiceMock)
+        {
+            _blockRepository = blockRepository;
+            _followRepository = followRepository;
+            _userServiceMock = userServiceMock;
         }
 
         FollowRepository getFollowRepository()
@@ -27,7 +37,7 @@ namespace UBB_SE_2024_Gaborment.Server.Relationships.Follow
 
         public void createFollow(string sender, string receiver)
         {
-            if (!(_blockRepository.GetBlocksBySender(sender).Any(b => b.getReceiver() == receiver) || _blockRepository.GetBlocksOfReceiver(receiver).Any(b => b.getSender() == sender)))
+            if (!(_blockRepository.GetBlocksBySender(sender).Any(b => b.getReceiver() == receiver) || _blockRepository.GetBlocksOfReceiver(receiver).Any(b => b.getSender() == sender) || _followRepository.GetFollowersOf(sender).Any(f => f.getReceiver() == receiver)))
             {
                 Follow followToBeAdded = new Follow(sender, receiver);
                 _followRepository.AddFollow(followToBeAdded);
@@ -35,7 +45,10 @@ namespace UBB_SE_2024_Gaborment.Server.Relationships.Follow
         }
         public void removeFollow(string sender, string receiver)
         {
-            _followRepository.RemoveFollow(sender, receiver);
+            if (!(_blockRepository.GetBlocksBySender(sender).Any(b => b.getReceiver() == receiver) || _blockRepository.GetBlocksOfReceiver(receiver).Any(b => b.getSender() == sender)))
+            {
+                _followRepository.RemoveFollow(sender, receiver);
+            }
         }
 
         public void updateCloseFriendStatus(string sender, string receiver)
@@ -54,7 +67,7 @@ namespace UBB_SE_2024_Gaborment.Server.Relationships.Follow
                 }
             }
         }
-
+        
         public List<Follow> getFollowersOf(string sender)
         {
             return _followRepository.GetFollowersOf(sender);
@@ -78,6 +91,7 @@ namespace UBB_SE_2024_Gaborment.Server.Relationships.Follow
             return _followRepository.GetFollowingOf(receiver).Select(follower => follower.getSender()).ToList();
         }
 
+        
         public Dictionary<string, List<Follow>> getAllFollowers()
         {
             return _followRepository.GetFollowers()
@@ -86,5 +100,44 @@ namespace UBB_SE_2024_Gaborment.Server.Relationships.Follow
         }
 
         //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+        public List<UserMock> getFollowersOfAsUserList(string sender)
+        {
+            List<string> followList = _followRepository.GetFollowersOf(sender).Select(follower => follower.getReceiver()).ToList(); ;
+            List<UserMock> followListUser = new List<UserMock>();
+            foreach (string user in followList) 
+            {
+                UserMock FollowedUser = _userServiceMock.GetUserById(user);
+                followListUser.Add(FollowedUser);
+            }
+            return followListUser;
+        }
+
+        public List<UserMock> getCloseFriendsOfAsUserList(string sender)
+        {
+            List<string> ClosefollowList = _followRepository.GetFollowersOf(sender)
+                            .Where(follow => follow.getCloseFriendStatus() == true)
+                            .Select(follow => follow.getReceiver())
+                            .ToList();
+            List<UserMock> followListUser = new List<UserMock>();
+            foreach (string user in ClosefollowList)
+            {
+                UserMock FollowedUser = _userServiceMock.GetUserById(user);
+                followListUser.Add(FollowedUser);
+            }
+            return followListUser;
+        }
+
+        public List<UserMock> getFollowingOfAsUserList(string receiver)
+        {
+            List<string> followList = _followRepository.GetFollowingOf(receiver).Select(follower => follower.getSender()).ToList();
+            List<UserMock> followListUser = new List<UserMock>();
+            foreach (string user in followList)
+            {
+                UserMock FollowedUser = _userServiceMock.GetUserById(user);
+                followListUser.Add(FollowedUser);
+            }
+            return followListUser;
+        }
     }
 }
